Attempt to fix the Android build
A few things needed to happen:

1. Replace `com.moowork.node` with `com.github.node-gradle.node`. The former is no longer maintained and wasn't working well anymore since we've upgraded both Node and Yarn in the past year.
2. Change the `type` of some of the newly added Android tokens.

We should eventually rethink the `types` property to see if we can come up with types that make more sense. The existing ones were borrowed from Style Dictionary, Amazon's open-source design tokens project. Speaking of Style Dictionary, it may make sense to re-evaluate it as we work
on semantic tokens in 2023.

I tested that this works by running `./gradlew build`. It fails on the `master` branch but passes on this commit.